Deleted unneeded exceptions/handlers/responses
