+++ src/main/java/voting/service/ParsingServiceImpl.java
@@ -8,9 +8,6 @@ import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
 import org.springframework.validation.*;
 import voting.dto.CandidateData;
-import voting.exception.CsvMultiErrorsException;
-import voting.exception.DTOMultiFieldsErrorsException;
-import voting.exception.DTOMultiObjectsErrorsException;
 
 import javax.validation.ConstraintViolation;
 import javax.validation.Validation;
