+++ src/main/java/voting/service/PartyServiceImpl.java
@@ -9,9 +9,6 @@ import org.springframework.validation.*;
 import org.springframework.web.multipart.MultipartFile;
 import voting.dto.CandidateData;
 import voting.dto.PartyData;
-import voting.exception.CsvMultiErrorsException;
-import voting.exception.DTOMultiFieldsErrorsException;
-import voting.exception.DTOMultiObjectsErrorsException;
 import voting.exception.NotFoundException;
 import voting.model.Candidate;
 import voting.model.Party;
